+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- holderId attribute to MerchantSession and MerchantPurchase
 
 ## [2.24.0] - 2026-02-12
 ### Added